move stuff to systeminformation.js
