+++ tests/unit/classes/systeminformation_spec.js
@@ -1,0 +1,7 @@
+const SystemInformation = require("../../../js/systeminformation");
+
+describe("SystemInformation", () => {
+	it("should output system information", async () => {
+		await expect(SystemInformation()).resolves.toContain("platform: linux");
+	});
+});
